TaggedMetricRegistry composability
This is a very quick thing to throw up, but basically internally
we have a bunch of producers of the same set of metrics, and we'd
like to be able to differentiate them (with tags).

With this PR, we'd be able to have multiple metric registries,
and merge them into a single metric registry.

Please don't merge, since there are a few things I need to think
through, but wanted to make sure that this doesn't sound super
unreasonable.