Optimize common logging paths where invocation logging is typically only enabled at TRACE or DEBUG
Use if/else as it generates more compact bytecode to keep the common
case of TRACE and DEBUG below the default JIT method threshold of 35
bytes.

Move uncommon case of invocation logging at INFO, WARN, or ERROR levels
to separate smaller method that is also below the JIT method threshold.