[break] Only one 'default' TaggedMetricRegistry (#249)
In the old world, libraries called `DefaultTaggedMetricRegistry.getDefault()` and everything was kinda OK.

However, we recently added the `SlidingWindowTaggedMetricRegistry`, which also has a `getDefault()` method.  I think this method is dangerous because there's now a chance some libraries will write to the first and other libraries write to the second. In WC we just emit metrics from _one_ registry, so this could lead to some metrics not being emitted!

- [atlasdb](https://github.com/palantir/atlasdb/blob/3b3031b86c972baacccc03d08ed9c40f8ec678a4/atlasdb-config/src/main/java/com/palantir/atlasdb/factory/TransactionManagers.java#L241)
- [conjure-java-runtime](https://github.com/palantir/conjure-java-runtime/blob/develop/okhttp-clients/src/main/java/com/palantir/conjure/java/okhttp/OkHttpClients.java#L94)

Obviously it's preferable for libraries to just ask users to provide a specific TaggedMetricRegistry, as this is much easier to test and swap in/out implementations! In the meantime, my proposal is to just give people one default to rule them all.