Fix CompositeInvocationEventHandler when handlers are disabled in flight
InvocationEventHandler implementations are often responsible for
setting up and tearing down thread state. When a handler is
enabled at the beginning of a call, and disabled at the end, it
should still be executed to clean up what it started. This matches
expectations when only a single handler is used.