Tritium instrumentation handles null InvocationContexts
A behavior change was introduced in #379 which resulted in handlers
which return a null InvocationContext being considered disabled,
which prevented the `onSuccess` and `onFailure` methods from
being invoked at all.

This fixes the null case by introducing internal sentinel values
to describe disabled handlers instead of overloading the meaning
of `null`.