Implement equals and hashcode on instrumented TLS components
Some libraries pool connections with a check for SSLSocketFactory
equality. If we provide two connections from the same factory
wrapped with the same name in different places. While it's
safer to pool connections from a constant factory object, not
sharing a pool between client instances, we shouldn't
unexpectably break this case.