Implement a lock-free exponentialy decaying reservoir
This solves both the nanotime issue reported in
https://github.com/dropwizard/metrics/pull/1654 as well
as correctness issues when the system clock changes in the original
implementation as this one only uses the precise clock.

More importantly, unexpected errors (OME, ThreadDeath, etc) cannot leave
this reservoir in a blocked state because it has no locks. If a thread
is descheduled while rescaling, another thread may complete the task
first.

This Reservoir allows us to specify a rescale threshold rather than
the hard coded hour.