LoggingInvocationEventHandler pre-computes level information
This avoids branching duplicating the level checks for is-enabled
and the logging call on each method invocation.

This commit originally used a new LoggingLevel.Visitor, however
upon inspection the visitor was unnecessary -- logging level is
a constant so we can do all branching up front. I've not rerun
the benchmarks yet, it's possible the lambda incurs an additional
pointer jump, however I'd expect it to be quickly inlined.