Tritium SSLEngine handshake instrumentation avoids reflection
Checking for older java 8 releases was necessary when ALPN support
was initially backported, however at this point it's incredibly
dangerous to run outdated runtimes, and our environments upgraded
long ago. Furthermore, the risk of introducing bugs when using
an old JRE is incredibly low, and requires a framework which
attempts to reflectively discover SSLEngine methods.