Tritium provides an annotation processor alternative to dynamic proxies
This makes the implementation more obvious, and moves work to build time
rather than runtime bytecode generation.

Drawbacks:
Unlike the runtime implementations, this solution does not support
adding arbitrary interfaces from the implementation to the instrumented
object. This may make implementations more obvious, but means that
there are cases (e.g. dependency injection) in which the annotation
processor does not produce behavior identical to the proxy.