Implement a dynamic-proxy annotation processor
This is nearly identical to the tritium processor, except it supports
the full breadth of `InvocationHandler` rather than tritium-specific
types. This moves proxy class creation from runtime to build time,
and works with graalvm native-images without any custom registration.

I'm not sure the tritium processor is the best places for this to
live. It's not a bad fit, and has a great deal of overlap with the
existing annotation processor.

I'm skeptical that `InvocationHandler` drop-in replacement is
the ideal solution. It will certainly unblock us with
minimal code churn, however in a perfect world we would provide
a clean API that can encapsulate the delegate invocation cleanly,
and provide metadata without necessarily exposing reflective types.
Unfortunately it would be time consuming to get such an API right,
and it's unlikely we'd get everything correct on the first attempt.
In fact it would likely involve several APIs tailored for specific use
cases...