Instrumented executors report the correct 'submitted' value after rejection
Previoulsy the 'submitted' metric was updated prior to interactions
with the delegate executor, producing data which suggested the
underlying thread pool or queue may be much larger than they are
in practice.

Now the meter is updated after tasks are submitted. The data may
be skewed/delayed in the case of a direct (same-thread) implementation,
however those are much rarer than thread pool shutdown, and generally
aren't instrumented.