Queued duration timer is only updated beyond a threshold
The queued duration estimate is only updated for tasks which
take longer than 250ms to begin executing. Not all tasks are
queued, so this should produce more realistic values in most
cases. However, it is possible that certain workloads do in
fact queue with very small durations, and we won't receive
that precise data anymore.